+++ src/main/java/com/anipick/backend/user/service/UserService.java
@@ -5,11 +5,11 @@ import com.anipick.backend.common.exception.CustomException;
 import com.anipick.backend.common.exception.ErrorCode;
 import com.anipick.backend.token.dto.TokenResponse;
 import com.anipick.backend.token.service.TokenService;
-import com.anipick.backend.user.controller.dto.LoginRequest;
+import com.anipick.backend.user.dto.LoginRequest;
 import com.anipick.backend.user.domain.LoginFormat;
 import com.anipick.backend.user.domain.User;
 import com.anipick.backend.user.domain.UserDefaults;
-import com.anipick.backend.user.controller.dto.SignUpRequest;
+import com.anipick.backend.user.dto.SignUpRequest;
 import com.anipick.backend.user.mapper.UserMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
